better tests

+++ rootfs/init/run.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 #
 
 #set -eo pipefail
@@ -6,7 +6,7 @@
 #set -u
 #set -x
 
-SERVER_CNF="/etc/mysql/my.cnf"
+# SERVER_CNF="/etc/mysql/my.cnf"
 SERVER_CNF_D="/etc/mysql"
 
 . /init/output.sh
@@ -18,19 +18,13 @@ config_check() {
 
   if [ -d /etc/my.cnf.d ]
   then
-    SERVER_CNF=/etc/my.cnf
+    # SERVER_CNF=/etc/my.cnf
     SERVER_CNF_D=/etc/my.cnf.d
     SERVER_CONFIG_CNF="${SERVER_CNF_D}/mariadb-server.cnf"
 
-    if [ -f "${SERVER_CNF_D}/mariadb-server.cnf" ]
-    then
-      rm -f "${SERVER_CNF_D}/mariadb-server.cnf"
-    fi
+    [ -f "${SERVER_CONFIG_CNF}" ] && rm -f "${SERVER_CONFIG_CNF}"
+    [ -d /etc/mysql ] && rm -rf /etc/mysql
 
-    if [ -d /etc/mysql ]
-    then
-      rm -rf /etc/mysql
-    fi
   fi
 }
 
@@ -82,17 +76,37 @@ bootstrap_database() {
 
     [ -f /root/.my.cnf ] && rm /root/.my.cnf
 
-    log_info "install initial databases"
-    mysql_install_db \
-      --user="${MARIADB_SYSTEM_USER}" \
-      > /dev/null 2> /dev/null
-    [ $? -gt 0 ] && exit $?
+    log_info "  install initial databases"
+    data=$(mysql_install_db \
+      --user="${MARIADB_SYSTEM_USER}")
 
-    log_info "start initial instance in safe mode to set passwords"
-    /usr/bin/mysqld_safe \
-      --syslog-tag=init \
-      > /dev/null 2> /dev/null &
-    [ $? -gt 0 ] && exit $?
+    result=${?}
+
+    if [ ${result} -gt 0 ]
+    then
+      while IFS= read -r line
+      do
+        log_error "${line}"
+      done < <(printf '%s\n' "${data}")
+
+      exit ${result}
+    fi
+
+    log_info "  start initial instance in safe mode to set passwords"
+    data=$(/usr/bin/mysqld_safe \
+      --syslog-tag=init &)
+
+    result=${?}
+
+    if [ ${result} -gt 0 ]
+    then
+      while IFS= read -r line
+      do
+        log_error "${line}"
+      done < <(printf '%s\n' "${data}")
+
+      exit ${result}
+    fi
 
     set +e
     retry=30
@@ -104,17 +118,17 @@ bootstrap_database() {
       #
       status=$(nc -v -w1 -z 127.0.0.1 3306 2>&1)
 
-      if [ $? -eq 0 ] && [ $(echo "${status}" | grep -c open) -eq 1 ]
+      if [ $? -eq 0 ] && [ "$(echo "${status}" | grep -c open)" -eq 1 ]
       then
         break
       else
-        sleep 5s
+        sleep 8s
         retry=$((retry - 1))
       fi
     done
     set -e
 
-    log_info "create privileges for root access"
+    log_info "  create privileges for root access"
     (
       echo "USE mysql;"
       echo "UPDATE user SET password = PASSWORD('${MARIADB_ROOT_PASSWORD}') WHERE user = 'root';"
@@ -123,11 +137,12 @@ bootstrap_database() {
       echo "GRANT ALL PRIVILEGES ON *.* TO 'root'@'localhost' WITH GRANT OPTION;"
       echo "FLUSH PRIVILEGES;"
     ) | mysql --host=localhost > /dev/null 2> /dev/null
+
     [ $? -gt 0 ] && exit $?
 
     sleep 2s
 
-    log_info "kill bootstrapping instance"
+    log_info "  kill bootstrapping instance"
     killall mysqld
     [ $? -gt 0 ] && exit $?
